rename charts folder/files
