+++ docker-compose.yml
@@ -30,6 +30,6 @@ services:
       - "5173:5173"
     volumes:
       - ./frontend/src:/app/src
-      - ./frontend/charts:/app/charts
+      - ./frontend/data:/app/data
     depends_on:
       - api
