+++ frontend/data/areaChart.ts
