+++ frontend/data/barChart.ts
